Command and Conquer Game. Grade 12 final project
For my final project for grade 12, I decided to make a game similar to Command and conquer. I am not finished yet. I plan on finishing it by the end of the semester (The start of February). By the end there will be a computer that has a base and the game will be cleaner and have more units and buildings.